+++ src/InputWithOptions/InputWithOptions.spec.js
@@ -276,12 +276,12 @@ const runInputWithOptionsTest = driverFactory => {
       expect(onSelect).toBeCalledWith(options[0]);
     });
 
-    it('should not call onSelect when a selected option is pressed', () => {
+    it('should call onSelect when a selected option is pressed', () => {
       const onSelect = jest.fn();
       const {driver, dropdownLayoutDriver} = createDriver(<InputWithOptions options={options} onSelect={onSelect} selectedId={options[0].id}/>);
       driver.focus();
       dropdownLayoutDriver.clickAtOption(0);
-      expect(onSelect).not.toBeCalled();
+      expect(onSelect).toBeCalled();
     });
 
     it('should call onFocus', () => {