+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Types of changes:
 - `<SideMenu/>` - support `dataHook` prop [#2256](https://github.com/wix/wix-style-react/pull/2256)
 - Polyfills - fix `raf` not setting all methods on the global variable [#2258](https://github.com/wix/wix-style-react/pull/2258)
 - `Tooltip` - remove `stopPropagation` from `onClick` as it breaks some use cases [#2260](https://github.com/wix/wix-style-react/pull/2260)
+- `<InputWithOptions/>` - fix `onSelect` to be called also when re-selecting same option [#2265](https://github.com/wix/wix-style-react/pull/2265/files)
 
 ## 5.3.2 - 2018-09-20